+++ game/scrabble.py
@@ -2,7 +2,6 @@ from game.board import Board
 from game.player import Player
 from game.models import BagTiles, Tile
 from game.dictionary import validate_word as dict
-import collections
 
 class ScrabbleGame:
     def __init__(self, playerCount: int):
@@ -10,11 +9,14 @@ class ScrabbleGame:
         self.bagTiles = BagTiles()
         self.players:list[Player] = []
         for index in range(playerCount):
-            self.players.append(Player())
+            self.players.append(Player(self.bagTiles, id=index+1))
         self.current_player = None
 
     def validateTurn(self, word, location, orientation):
-        return self.putWord(word, location, orientation)
+        while len(self.bagTiles) > 0:
+            self.putWord(word, location, orientation)
+        else:
+            return False;
 
     def next_turn(self):
         if self.current_player is None:
@@ -26,18 +28,10 @@ class ScrabbleGame:
             self.current_player = self.players[index]
 
     def validateWord(self, word, location, orientation):
-        """ 
-        if self.current_player.hasWord(word) and self.board.validate_word_inside_board(word, location, orientation):
-        """
         if self.current_player.hasWord(word) and self.board.validate_word_inside_board(word, location, orientation) and dict(word):
             return True;
         else:
             return False;
-    def get_words():
-        '''
-        Obtener las posibles palabras que se pueden formar, dada una palabra, ubicacion y orientacion 
-        Preguntar al usuario, por cada una de esas palabras, las que considera reales
-        '''
     
     def putWord(self, word, location, orientation):
         (x, y) = location