+++ tests/test_player.py
@@ -1,14 +1,70 @@
 import unittest
 from game.player import Player
+from game.scrabble import ScrabbleGame
 from game.models import Tile, BagTiles
 
 class TestPlayer(unittest.TestCase):
     def test_init(self):
-        player1 = Player()
+        bagTiles = BagTiles()
+        player1 = Player(bagTiles, id=1)
         self.assertEqual(len(player1.tiles), 7)
 
+    def test_getTiles(self):
+        bagTiles = BagTiles()
+        Player(bagTiles, id=1)
+        self.assertEqual(len(bagTiles.tiles), 94)
+
+    def test_fillTiles(self):
+        scrabbleGame = ScrabbleGame(2)
+        scrabbleGame.current_player = scrabbleGame.players[1]
+        scrabbleGame.current_player.tiles = [
+            Tile('A', 1), 
+            Tile('A', 1), 
+            Tile('C', 1), 
+            Tile('X', 1), 
+            Tile('S', 1), 
+            Tile('F', 1), 
+            Tile('I', 1)
+        ]
+        scrabbleGame.putWord("casa", (3, 2), "V")
+        self.assertEqual(len(scrabbleGame.current_player.tiles), 3)
+        scrabbleGame.current_player.fillTiles(scrabbleGame.bagTiles)
+        self.assertEqual(len(scrabbleGame.bagTiles.tiles), 83)
+        self.assertEqual(len(scrabbleGame.current_player.tiles), 7)
+
+    def test_fillTilesWithDoubleTurn(self):
+        scrabbleGame = ScrabbleGame(2)
+        scrabbleGame.current_player = scrabbleGame.players[1]
+        scrabbleGame.current_player.tiles = [
+            Tile('A', 1), 
+            Tile('A', 1), 
+            Tile('C', 1), 
+            Tile('X', 1), 
+            Tile('S', 1), 
+            Tile('F', 1), 
+            Tile('I', 1)
+        ]
+        scrabbleGame.putWord("casa", (3, 2), "V")
+        scrabbleGame.current_player.fillTiles(scrabbleGame.bagTiles)
+        scrabbleGame.current_player = scrabbleGame.players[0]
+        scrabbleGame.current_player.tiles = [
+            Tile('X', 1), 
+            Tile('S', 1), 
+            Tile('I', 1), 
+            Tile('U', 1), 
+            Tile('A', 1), 
+            Tile('G', 1), 
+            Tile('I', 1)
+        ]
+        scrabbleGame.putWord("guias", (7, 3), "H")
+        self.assertEqual(len(scrabbleGame.current_player.tiles), 2)
+        scrabbleGame.current_player.fillTiles(scrabbleGame.bagTiles)
+        self.assertEqual(len(scrabbleGame.bagTiles.tiles), 78)
+        self.assertEqual(len(scrabbleGame.current_player.tiles), 7)
+        
     def test_playerHasWord(self):
-        player1 = Player()
+        bagTiles = BagTiles()
+        player1 = Player(bagTiles, id=1)
         player1.tiles = [
             Tile('A', 1), 
             Tile('A', 1), 
@@ -21,7 +77,8 @@ class TestPlayer(unittest.TestCase):
         self.assertEqual(player1.hasWord("casa"), True)
 
     def test_playerDoesntHaveWord(self):
-        player1 = Player()
+        bagTiles = BagTiles()
+        player1 = Player(bagTiles, id=1)
         player1.tiles = [
             Tile('T', 1), 
             Tile('C', 1), 
@@ -34,7 +91,8 @@ class TestPlayer(unittest.TestCase):
         self.assertEqual(player1.hasWord("torpe"), False)
     
     def test_playerAlmostHasWord(self):
-        player1 = Player()
+        bagTiles = BagTiles()
+        player1 = Player(bagTiles, id=1)
         player1.tiles = [
             Tile('T', 1), 
             Tile('C', 1), 
@@ -47,7 +105,8 @@ class TestPlayer(unittest.TestCase):
         self.assertEqual(player1.hasWord("torpe"), False)
 
     def test_playerHasWordWithMoreTiles(self):
-        player1 = Player()
+        bagTiles = BagTiles()
+        player1 = Player(bagTiles, id=1)
         player1.tiles = [
             Tile('A', 1), 
             Tile('A', 1), 
@@ -58,15 +117,11 @@ class TestPlayer(unittest.TestCase):
             Tile('I', 1)
         ]
         self.assertEqual(player1.hasWord("casa"), True)
-    
-    def test_playerHasNoTiles(self):
-        player1 = Player()
-        self.assertEqual(player1.hasWord("torpe"), False)
 
     def test_validate_user_has_letters(self):
-        player1 = Player()
-        bagTile = BagTiles()
-        bagTile.tiles = [
+        bagTiles = BagTiles()
+        player1 = Player(bagTiles, id=1)
+        bagTiles.tiles = [
             Tile('O', 1),
             Tile('L', 1),
             Tile('H', 1),
@@ -75,7 +130,7 @@ class TestPlayer(unittest.TestCase):
             Tile('U', 1),
             Tile('M', 1)
         ]
-        player1.tiles.extend(bagTile.tiles)
+        player1.tiles.extend(bagTiles.tiles)
         tiles = [
             Tile('H', 1),
             Tile('O', 1),
@@ -86,9 +141,9 @@ class TestPlayer(unittest.TestCase):
         self.assertEqual(is_valid, True)
 
     def test_user_has_letters2(self):
-        player1 = Player()
-        bagTile = BagTiles()
-        bagTile.tiles = [
+        bagTiles = BagTiles()
+        player1 = Player(bagTiles, id=1)
+        bagTiles.tiles = [
             Tile('O', 1),
             Tile('A', 1),
             Tile('F', 1),
@@ -97,7 +152,7 @@ class TestPlayer(unittest.TestCase):
             Tile('I', 1),
             Tile('N', 1)
         ]
-        player1.tiles.extend(bagTile.tiles)
+        player1.tiles.extend(bagTiles.tiles)
         tiles = [
             Tile('A', 1),
             Tile('V', 1),